+++ src/mongo/db/index/index_build_interceptor.cpp
@@ -254,7 +254,7 @@ Status IndexBuildInterceptor::drainWritesIntoIndex(OperationContext* opCtx,
 
     int logLevel = (_numApplied - appliedAtStart > 0) ? 0 : 1;
     LOGV2_DEBUG(20689,
-                logLevel.toInt(),
+                logLevel,
                 "index build: drained side writes",
                 "numApplied"_attr = (_numApplied - appliedAtStart),
                 "totalInserted"_attr = totalInserted,
