+++ src/mongo/db/repl/replication_coordinator_impl_heartbeat.cpp
@@ -962,7 +962,7 @@ void ReplicationCoordinatorImpl::_cancelAndRescheduleElectionTimeout_inlock() {
     }
     if (wasActive) {
         LOGV2_FOR_ELECTION(4615649,
-                           cancelAndRescheduleLogLevel.toInt(),
+                           cancelAndRescheduleLogLevel,
                            "Canceling election timeout callback at {when}",
                            "Canceling election timeout callback",
                            "when"_attr = _handleElectionTimeoutWhen);
@@ -980,7 +980,7 @@ void ReplicationCoordinatorImpl::_cancelAndRescheduleElectionTimeout_inlock() {
     if (wasActive) {
         // The log level here is 4 once per second, otherwise 5.
         LOGV2_FOR_ELECTION(4615650,
-                           cancelAndRescheduleLogLevel.toInt(),
+                           cancelAndRescheduleLogLevel,
                            "Rescheduling election timeout callback at {when}",
                            "Rescheduling election timeout callback",
                            "when"_attr = when);
