+++ src/mongo/executor/connection_pool.cpp
@@ -645,7 +645,7 @@ Future<ConnectionPool::ConnectionHandle> ConnectionPool::SpecificPool::getConnec
         if (conn) {
             LOGV2_DEBUG(
                 22559,
-                kDiagnosticLogLevel.toInt(),
+                kDiagnosticLogLevel,
                 "Requesting new connection to {hostAndPort}--using existing idle connection",
                 "hostAndPort"_attr = _hostAndPort);
             return Future<ConnectionPool::ConnectionHandle>::makeReady(std::move(conn));
@@ -657,7 +657,7 @@ Future<ConnectionPool::ConnectionHandle> ConnectionPool::SpecificPool::getConnec
         timeout = pendingTimeout;
     }
     LOGV2_DEBUG(22560,
-                kDiagnosticLogLevel.toInt(),
+                kDiagnosticLogLevel,
                 "Requesting new connection to {hostAndPort} with timeout {timeout}",
                 "hostAndPort"_attr = _hostAndPort,
                 "timeout"_attr = timeout);
@@ -728,7 +728,7 @@ void ConnectionPool::SpecificPool::finishRefresh(ConnectionInterface* connPtr, S
     // to our internal one.
     if (status.code() == ErrorCodes::NetworkInterfaceExceededTimeLimit) {
         LOGV2_DEBUG(22562,
-                    kDiagnosticLogLevel.toInt(),
+                    kDiagnosticLogLevel,
                     "Pending connection to host {hostAndPort} did not complete within the "
                     "connection timeout, retrying with a new connection;{openConnections} "
                     "connections to that host remain open",
@@ -740,7 +740,7 @@ void ConnectionPool::SpecificPool::finishRefresh(ConnectionInterface* connPtr, S
     // Pass a failure on through
     if (!status.isOK()) {
         LOGV2_DEBUG(22563,
-                    kDiagnosticLogLevel.toInt(),
+                    kDiagnosticLogLevel,
                     "Connection failed to {hostAndPort} due to {status}",
                     "hostAndPort"_attr = _hostAndPort,
                     "status"_attr = redact(status));
@@ -751,14 +751,14 @@ void ConnectionPool::SpecificPool::finishRefresh(ConnectionInterface* connPtr, S
     // If the host and port were dropped, let this lapse and spawn new connections
     if (!conn || conn->getGeneration() != _generation) {
         LOGV2_DEBUG(22564,
-                    kDiagnosticLogLevel.toInt(),
+                    kDiagnosticLogLevel,
                     "Dropping late refreshed connection to {hostAndPort}",
                     "hostAndPort"_attr = _hostAndPort);
         return;
     }
 
     LOGV2_DEBUG(22565,
-                kDiagnosticLogLevel.toInt(),
+                kDiagnosticLogLevel,
                 "Finishing connection refresh for {hostAndPort}",
                 "hostAndPort"_attr = _hostAndPort);
 
@@ -814,7 +814,7 @@ void ConnectionPool::SpecificPool::returnConnection(ConnectionInterface* connPtr
         _processingPool[connPtr] = std::move(conn);
 
         LOGV2_DEBUG(22568,
-                    kDiagnosticLogLevel.toInt(),
+                    kDiagnosticLogLevel,
                     "Refreshing connection to {hostAndPort}",
                     "hostAndPort"_attr = _hostAndPort);
         connPtr->refresh(_parent->_controller->pendingTimeout(),
@@ -827,7 +827,7 @@ void ConnectionPool::SpecificPool::returnConnection(ConnectionInterface* connPtr
 
     // If it's fine as it is, just put it in the ready queue
     LOGV2_DEBUG(22569,
-                kDiagnosticLogLevel.toInt(),
+                kDiagnosticLogLevel,
                 "Returning ready connection to {hostAndPort}",
                 "hostAndPort"_attr = _hostAndPort);
     addToReady(std::move(conn));
@@ -847,7 +847,7 @@ void ConnectionPool::SpecificPool::addToReady(OwnedConnection conn) {
     // returnConnection
     auto returnConnectionFunc = guardCallback([this, connPtr]() {
         LOGV2_DEBUG(22570,
-                    kDiagnosticLogLevel.toInt(),
+                    kDiagnosticLogLevel,
                     "Triggered refresh timeout for {hostAndPort}",
                     "hostAndPort"_attr = _hostAndPort);
         auto conn = takeFromPool(_readyPool, connPtr);
@@ -931,7 +931,7 @@ void ConnectionPool::SpecificPool::processFailure(const Status& status) {
     }
 
     LOGV2_DEBUG(22573,
-                kDiagnosticLogLevel.toInt(),
+                kDiagnosticLogLevel,
                 "Failing requests to {hostAndPort}",
                 "hostAndPort"_attr = _hostAndPort);
     _requests.clear();
@@ -972,7 +972,7 @@ void ConnectionPool::SpecificPool::spawnConnections() {
 
     if (_health.isFailed) {
         LOGV2_DEBUG(22574,
-                    kDiagnosticLogLevel.toInt(),
+                    kDiagnosticLogLevel,
                     "Pool for {hostAndPort} has failed recently. Postponing any attempts to spawn "
                     "connections",
                     "hostAndPort"_attr = _hostAndPort);
@@ -981,7 +981,7 @@ void ConnectionPool::SpecificPool::spawnConnections() {
 
     auto controls = _parent->_controller->getControls(_id);
     LOGV2_DEBUG(22575,
-                kDiagnosticLogLevel.toInt(),
+                kDiagnosticLogLevel,
                 "Comparing connection state for {hostAndPort} to Controls: {controls}",
                 "hostAndPort"_attr = _hostAndPort,
                 "controls"_attr = controls);
@@ -1005,7 +1005,7 @@ void ConnectionPool::SpecificPool::spawnConnections() {
     auto allowance = std::min(controls.targetConnections - totalConnections,
                               controls.maxPendingConnections - pendingConnections);
     LOGV2_DEBUG(22577,
-                kDiagnosticLogLevel.toInt(),
+                kDiagnosticLogLevel,
                 "Spawning {allowance} connections to {hostAndPort}",
                 "allowance"_attr = allowance,
                 "hostAndPort"_attr = _hostAndPort);
@@ -1136,7 +1136,7 @@ void ConnectionPool::SpecificPool::updateController() {
         inUseConnections(),
     };
     LOGV2_DEBUG(22578,
-                kDiagnosticLogLevel.toInt(),
+                kDiagnosticLogLevel,
                 "Updating controller for {hostAndPort} with State: {state}",
                 "hostAndPort"_attr = _hostAndPort,
                 "state"_attr = state);
@@ -1195,7 +1195,7 @@ void ConnectionPool::SpecificPool::updateState() {
     if (_health.isShutdown) {
         // If we're in shutdown, there is nothing to update. Our clients are all gone.
         LOGV2_DEBUG(22579,
-                    kDiagnosticLogLevel.toInt(),
+                    kDiagnosticLogLevel,
                     "{hostAndPort} is dead",
                     "hostAndPort"_attr = _hostAndPort);
         return;
