+++ src/mongo/s/catalog_cache.cpp
@@ -601,7 +601,7 @@ void CatalogCache::_scheduleDatabaseRefresh(WithLock lk,
             : 1;
         LOGV2_FOR_CATALOG_REFRESH(
             24101,
-            logLevel.toInt(),
+            logLevel,
             "Refreshed cached database entry for {db} to version {newDbVersion} from version "
             "{oldDbVersion}. Took {duration}",
             "Refreshed cached database entry",
@@ -712,7 +712,7 @@ void CatalogCache::_scheduleCollectionRefresh(WithLock lk,
                 : 1;
             LOGV2_FOR_CATALOG_REFRESH(
                 24104,
-                logLevel.toInt(),
+                logLevel,
                 "Refreshed cached collection {namespace} to version {newVersion} from version "
                 "{oldVersion}. Took {duration}",
                 "Refreshed cached collection",
